# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/createnewproductProizvoditel": {
            "post": {
                "tags": [
                    "AuthProducts"
                ],
                "summary": "createnewproductProizvoditel",
                "description": "createnewproductProizvoditel",
                "operationId": "createnewproductProizvoditel",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "NULL"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "1"
                                    },
                                    "category_name": {
                                        "type": "string",
                                        "format": "phone",
                                        "example": "Жилая мебель"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "phone",
                                        "example": "Furnitur"
                                    },
                                    "frame": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Корпус"
                                    },
                                    "facades": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Фасады"
                                    },
                                    "length": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12"
                                    },
                                    "height": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12"
                                    },
                                    "price": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "10000"
                                    },
                                    "tabletop": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Столешница"
                                    },
                                    "photo[]": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "photo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/GetAllProduct": {
            "get": {
                "tags": [
                    "AuthProducts"
                ],
                "summary": "Get all products paginate 10",
                "description": "Returns list of projects",
                "operationId": "GetAllProduct",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "api/deleteAuthUserProduct": {
            "post": {
                "tags": [
                    "AuthProducts"
                ],
                "summary": "deleteAuthUserProduct",
                "description": "deleteAuthUserProduct",
                "operationId": "deleteAuthUserProduct",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "NULL"
                                ],
                                "properties": {
                                    "product_id[]": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/GetAllProductOneUser": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get All ProductO ne User paginate 10",
                "description": "Returns list of projects",
                "operationId": "GetAllProductOneUser",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "api/GetcategoryOneuserprduct": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "GetcategoryOneuserprduct",
                "description": "GetcategoryOneuserprduct",
                "operationId": "GetcategoryOneuserprduct",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "NULL"
                                ],
                                "properties": {
                                    "category_name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "Кухни"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/getOneProizvoditel/user_id{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get All ProductO one User ",
                "description": "Returns list of projects",
                "operationId": "getOneProizvoditel",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "api/filtergetOneProizvoditel": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "filtergetOneProizvoditel",
                "description": "filtergetOneProizvoditel",
                "operationId": "filtergetOneProizvoditel",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "NULL"
                                ],
                                "properties": {
                                    "category_name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "Кухни"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}